importing essence files from bilal's breaking_conjure repo

+++ tests/parse_print/acceptOutput.sh
@@ -6,9 +6,12 @@ set -o nounset
 TESTCASE="$1"
 
 if [ -d "${TESTCASE}" ]; then
-    echo "Accepting the output of ${TESTCASE}"
-    touch "${TESTCASE}"/stdout "${TESTCASE}"/stderr
-    mv "${TESTCASE}"/stdout "${TESTCASE}"/stdout.expected
-    mv "${TESTCASE}"/stderr "${TESTCASE}"/stderr.expected
-    find "${TESTCASE}" -size 0 -delete
+    NUM=$(ls "${TESTCASE}/"*.essence 2> /dev/null | wc -l)
+    if [ "$NUM" -eq "1" ]; then
+        echo "Accepting the output of ${TESTCASE}"
+        touch "${TESTCASE}"/stdout "${TESTCASE}"/stderr
+        cp "${TESTCASE}"/stdout "${TESTCASE}"/stdout.expected
+        cp "${TESTCASE}"/stderr "${TESTCASE}"/stderr.expected
+        find "${TESTCASE}" -size 0 -delete
+    fi
 fi
